Исправляет и дополняет валидацию prop-types
